feat: updated Argument<T> type for functional compatibility with other architectures too
